re: 3021 exporting packages that reference original data model and icons issues
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@7917 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/java/org.wcs.smart.dataentry/src/org/wcs/smart/dataentry/DataentryHibernateManager.java
@@ -137,8 +137,8 @@ public class DataentryHibernateManager extends HibernateManager {
 	 * @param {@link ConfigurableModel} who's filestore to delete
 	 */
 	public static void deleteFilestore(ConfigurableModel model) {
-		Path fileStore = Paths.get(model.getFileDataStoreLocation());
-		if (Files.exists(fileStore)) {
+		Path fileStore =  model.getFileDataStoreLocation();
+		if (fileStore != null && Files.exists(fileStore)) {
 			try {
 				SmartUtils.deleteDirectory(fileStore);
 			} catch(Exception ex) {
